refactor(layering): close two backward edges by moving types down
The 07-29 audit's §11.1 table of backward edges is otherwise complete, but two
were missing from it. Both were a type in the higher layer appearing in the
lower layer's signatures — the same shape as the core <-> compute cycle R-11
fixed in #1207.

- compute -> exec: src/compute/weight_dispatch.h included exec/weight_handle.h
  because gemm_dispatch/gemv_dispatch/gemm_grouped_dispatch take
  `const WeightHandle&`. One backward edge against forty the other way.
  WeightHandle depends on nothing but core/, so it moved to
  src/core/weight_handle.h. WeightRegistry stays in exec/ — it is the
  executor's container, not a type the kernels need. Edge count now 0.
- quant -> compute: src/quant/mxfp4_gemm.h included
  compute/gemm_cutlass_mxfp4_sm120.h for CutlassMxFP4Weight, a POD with no
  project dependencies that describes a QUANTISED WEIGHT LAYOUT. Moved to
  src/quant/cutlass_mxfp4_weight.h, which is where it belongs by meaning as
  well as by layer.

Also the two compute -> model includes §11.1 itself called avoidable:
embedding.cu and gemm_dp4a.cu pulled the 800-LOC model/model_config.h with the
comment "// QType" while using neither ModelConfig nor FFNActivation. QType is
in core/qtype.h. compute -> model goes 9 -> 7; the remaining seven are real
(tokenizer for the four constrainers, model.h for encoder_forward, mtp_head,
FFNActivation).

NOT fixed, deliberately: one quant -> compute edge remains, and it is not an
include artefact. src/quant/nvfp4_gemm.cu calls gemm() because the NVFP4
fallback dequantises to FP16 and hands off to the dense GEMM. Inverting that
moves a dispatch decision on a GEMM path — not something to do unmeasured on a
repo whose gpu ctest label never runs.

Checked on the same sweep and NOT findings, recorded so the next pass skips
them: god-functions by callee count (led by forward_logits at 216, which the
ledger already records as intrinsically forward-pass-coupled) and repeated
function names (operator= in 27 files, init in 19, reset in 16 — noise).
kv_cache_dtype "in 4 representations" could not be confirmed either way.

make dev green, ctest -L unit 4/4, clang-format clean on changed lines,
anchor gate 65/65.
