+++ src/compute/gemm_cutlass_mxfp4_sm120.h
@@ -4,6 +4,8 @@
 #include <cstdint>
 #include <cstddef>
 
+#include "quant/cutlass_mxfp4_weight.h"  // CutlassMxFP4Weight (moved down, see there)
+
 namespace imp {
 
 // MXFP4 weight data for CUTLASS sm_120 block-scaled GEMM.
@@ -19,18 +21,6 @@ namespace imp {
 // For our use case: we convert existing NVFP4 weights (UE4M3 per 16)
 // to MXFP4 format (UE8M0 per 32) to use the MXFP4 tensor core path.
 // This may trade some precision for potentially better hardware scheduling.
-
-struct CutlassMxFP4Weight {
-    const void* data = nullptr;     // [N, K/2] packed E2M1 nibbles
-    void* scale_factors = nullptr;  // SfAtom layout UE8M0 (for CUTLASS prefill)
-    void* linear_scales = nullptr;  // [N, K/32] row-major UE8M0 (for GEMV decode)
-    float tensor_scale = 1.0f;      // deferred global scale
-    int64_t N = 0;
-    int64_t K = 0;
-    size_t sf_bytes = 0;
-    bool owns_data = false;  // true if data was allocated (Hadamard path)
-    int hadamard_bs = 0;     // Hadamard block size for online rotation (0=disabled)
-};
 
 // Compute SfAtom buffer size for MXFP4 (UE8M0, SFVecSize=32).
 size_t cutlass_mxfp4_sf_size(int rows, int K);
