+++ src/compute/gemm_dp4a.cu
@@ -1,7 +1,7 @@
 #include "compute/gemm.h"
 #include "compute/gemv_dp4a_traits.cuh"
 #include "runtime/pdl.h"
-#include "model/model_config.h"  // QType
+#include "core/qtype.h"
 #include "core/logging.h"
 
 #include <cuda_fp16.h>
